Build: Isolate platform plugins to an app specific directory
When building the stx-rook-ceph RPM for rook-ceph-apps the helm
plugins are installed in a location that could be populated with other
app plugins if their spec files are not properly set up.

Adjust the spec to provide an app specific location for the plugins to
ensure that no other app plugins are included in the application
tarball.

In addition, sync the version of app with the plugins so that debugging
of the application lifecycle operations in the logs are easier during
application updates.

Partial-Bug: #1920066
Signed-off-by: Robert Church <robert.church@windriver.com>
Change-Id: I774a31784450592e99caab3d991892642938f1ba