Update yaml structures to be compatible with k8s 1.23+
- Removed the "beta1" in apiextensios version
- Declare the field types in CustomResourceDefinition kinds.
The declaration is mandatory to apply the custom resources
in kubernetes.
- Update CRDs
- Update resources to use the new CRDs
- Add wait time for the ceph osd to create the pools. This is
necessary to generate the certificate keys properly.
- Increase number of restarts limit for job rook-ceph-provision
to give more time for the dependencies to get ready.
- Switch ceph-config-helper to ubuntu_bionic-20220802
- Update ceph to v16.2.9

Test Plan:
PASS: Apply rook-ceph-app with the changes in k8s v1.23
PASS: Apply rook-ceph-app with the changes in k8s v1.24

Story: 2009138
Task: 45647

Signed-off-by: Daian Cardoso Sganderlla <Daian.CardosoSganderlla@windriver.com>
Change-Id: Ifc1d905b49504c3ac8d1a1fd4dd3d6661661a4cd