Remove openstack-helm-infra build dependency
After [1] was merged, the build of this package started to fail as it
has a dependency on the `openstack-helm-infra` package, provided by the
`openstack-armada-app` repository. However, this dependency turned out
to be unnecessary for this app, as it's not being used anywhere.

This change removes the unnecessary openstack-helm-infra build
dependency for stx-rook-ceph build.

[1] <https://review.opendev.org/c/starlingx/manifest/+/893085>

Test Plan:
PASS: build-pkgs -c -p python3-k8sapp-rook,stx-rook-ceph

Closes-Bug: 2033668

Change-Id: I259ae358af0aedff3977beff8a4fc1debe78da4b
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>